Revert "Publish as ESM-only (#11)"
This reverts commit d52e67004b117ab0df949eb0308535a715a09d2d.
