--- .changeset/modern-kiwis-look.md
@@ -1,5 +1,0 @@
----
-'@as-integrations/express5': major
----
-
-Publish as ESM-only, and require Node v20.19 (which enables require(esm) support by default).
