+++ src/__tests__/expressSpecific.test.ts
@@ -3,9 +3,10 @@ import request from 'supertest';
 import compression, { filter as defaultFilter } from 'compression';
 import type { BaseContext } from '@apollo/server';
 import { ApolloServer } from '@apollo/server';
-import { expressMiddleware } from '../../dist/index.js';
+import { expressMiddleware } from '..';
 import { it, expect } from '@jest/globals';
 import cors from 'cors';
+import { resolvable } from './utils';
 
 it('gives helpful error if json middleware is not installed', async () => {
   const server = new ApolloServer({ typeDefs: 'type Query {f: ID}' });
@@ -202,36 +203,3 @@ it('supporting doubly-encoded variables example from migration guide', async () 
 
   await server.stop();
 });
-
-// Copyright 2019 Joseph Gentle
-
-// Permission to use, copy, modify, and / or distribute this software for any
-// purpose with or without fee is hereby granted, provided that the above
-// copyright notice and this permission notice appear in all copies.
-
-// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
-// REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND
-// FITNESS.IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
-//   INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
-// LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
-// OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
-// PERFORMANCE OF THIS SOFTWARE.
-
-export type Resolvable<T> = Promise<T> & {
-  resolve: (t: T) => void;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  reject: (e: any) => void;
-};
-
-export function resolvable<T = void>(): Resolvable<T> {
-  let resolve: (val: T) => void;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let reject: (err: any) => void;
-  const promise = new Promise<T>((_resolve, _reject) => {
-    resolve = _resolve;
-    reject = _reject;
-  }) as Resolvable<T>;
-  promise.resolve = resolve!;
-  promise.reject = reject!;
-  return promise;
-}
