+++ src/__tests__/utils.ts
@@ -1,0 +1,54 @@
+import type { Server } from 'http';
+import type { AddressInfo } from 'net';
+import { format } from 'url';
+
+export function urlForHttpServer(httpServer: Server): string {
+  const { address, port } = httpServer.address() as AddressInfo;
+
+  // Convert IPs which mean "any address" (IPv4 or IPv6) into localhost
+  // corresponding loopback ip. Note that the url field we're setting is
+  // primarily for consumption by our test suite. If this heuristic is wrong for
+  // your use case, explicitly specify a frontend host (in the `host` option
+  // when listening).
+  const hostname = address === '' || address === '::' ? 'localhost' : address;
+
+  return format({
+    protocol: 'http',
+    hostname,
+    port,
+    pathname: '/',
+  });
+}
+
+// Copyright 2019 Joseph Gentle
+
+// Permission to use, copy, modify, and / or distribute this software for any
+// purpose with or without fee is hereby granted, provided that the above
+// copyright notice and this permission notice appear in all copies.
+
+// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+// REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND
+// FITNESS.IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+//   INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
+// LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
+// OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
+// PERFORMANCE OF THIS SOFTWARE.
+
+export type Resolvable<T> = Promise<T> & {
+  resolve: (t: T) => void;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  reject: (e: any) => void;
+};
+
+export function resolvable<T = void>(): Resolvable<T> {
+  let resolve: (val: T) => void;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  let reject: (err: any) => void;
+  const promise = new Promise<T>((_resolve, _reject) => {
+    resolve = _resolve;
+    reject = _reject;
+  }) as Resolvable<T>;
+  promise.resolve = resolve!;
+  promise.reject = reject!;
+  return promise;
+}
